pyproject.toml
[tool.poetry]
name = "chibawest-gamecenter-bot"
version = "0.1.0"
description = ""
authors = ["mytk0u0 <mstk344+mytk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.18.0"
discord = "^1.7.3"
google-cloud-secret-manager = "^2.7.0"

[tool.poetry.dev-dependencies]
ipython = "^7.26.0"
black = "^21.7b0"
flake8 = "^3.9.2"
mypy = "^0.910"

[tool.poetry.scripts]
start = 'chibawest_gamecenter_bot.main:run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
